+++ src/pages/Users.tsx
@@ -1,52 +1,39 @@
 import { FunctionComponent, SyntheticEvent, useEffect, useState } from 'react';
 
-import { DataGrid, GridRowsProp, GridColDef, ptBR, GridRenderCellParams } from '@mui/x-data-grid';
+import { DataGrid, GridRowsProp, GridColDef, GridRenderCellParams } from '@mui/x-data-grid';
 import { Close, HowToReg } from '@mui/icons-material';
-import { useQuery } from '@tanstack/react-query';
+import { useQuery, useQueryClient } from '@tanstack/react-query';
 
 import ConfirmDialog from '../components/ConfirmDialog';
 import SnackbarAlert from '../components/SnackbarAlert';
-import DataGridOverlay from '../components/DataGrid/DataGridOverlay';
-import DataGridToolbar from '../components/DataGrid/DataGridToolbar';
-import DataGridLoading from '../components/DataGrid/DataGridLoading';
-import { createdAtColumnType, deleteColumnType } from '../components/DataGrid/DataGridCustomColumn';
+import { dataGridBasePropDefinitions } from '../components/DataGrid/DataGridBaseConfig';
+import { createdAtColumnType, deleteColumnType } from '../components/DataGrid/DataGridCustomColumns';
 
-import { api, IBaseResponse, PaginatedResponseType, queryClient } from '../services/api';
-import { UserRoleType, UserType } from '../services/user';
-import { UserAuth } from '../context/AuthProvider';
+import { api, PaginatedResponseType } from '../services/api';
+import { UserRoleType, UserType } from '../models/user';
+import { useAuth } from '../context/AuthProvider';
+import { getUsers } from '../services/user';
 
 interface UsersProps {}
 
 const Users: FunctionComponent<UsersProps> = () => {
   const [uid, setUid] = useState('');
   const [page, setPage] = useState<number>(0);
-  const [count, setCount] = useState<number>(0);
   const [pageSize, setPageSize] = useState<number>(10);
   const [confirmDelete, setConfirmDelete] = useState(false);
-  const [rows, setRows] = useState<GridRowsProp<UserType>>([]);
-  const [rowCountState, setRowCountState] = useState<number>(count);
+  const [rowCountState, setRowCountState] = useState<number>(0);
+  const [rowsState, setRowsState] = useState<GridRowsProp<UserType>>([]);
   const [showSuccessDeleteMessage, setShowSuccessDeleteMessage] = useState(false);
 
-  const { user } = UserAuth();
-  const localToken = sessionStorage.getItem('accessToken');
+  const { user } = useAuth();
+  const accessToken = user?.accessToken || sessionStorage.getItem('accessToken');
+  const queryClient = useQueryClient();
 
-  const fetchUsers = async (page: number, pageSize: number): Promise<PaginatedResponseType<UserType>> => {
-    const accessToken = user?.accessToken || localToken;
-    const { data: response }: IBaseResponse = await api.get(`/user?page=${page + 1}&pageSize=${pageSize}`, {
-      headers: { Authorization: `Bearer ${accessToken}` },
-    });
-    return response.data;
-  };
-
-  const {
-    isFetching,
-    isError,
-    refetch,
-    data: usersList,
-  } = useQuery<PaginatedResponseType<UserType>>(['usersList', page, pageSize], async () => fetchUsers(page, pageSize), {
-    keepPreviousData: true,
-    staleTime: 2000 * 60, // 2 minutes
-  });
+  const { isLoading, isFetching, isError, data } = useQuery<PaginatedResponseType<UserType>>(
+    ['usersList', page, pageSize],
+    () => getUsers(page, pageSize, { accessToken }),
+    { keepPreviousData: true, staleTime: 2000 * 60 },
+  );
 
   const handleSuccessDeleteClose = (_event?: SyntheticEvent | Event, reason?: string) => {
     if (reason === 'clickaway') return;
@@ -54,33 +41,20 @@ const Users: FunctionComponent<UsersProps> = () => {
   };
 
   const handleDelete = async (confirm = false) => {
-    const accessToken = user?.accessToken || localToken;
-
     setConfirmDelete(false);
-
     if (!confirm) return;
-
-    await api.delete(`/user/${uid}`, {
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-      },
-    });
-
+    await api.delete('/user/' + uid, { headers: { Authorization: `Bearer ${accessToken}` } });
     queryClient.invalidateQueries(['usersList']);
     setShowSuccessDeleteMessage(true);
-    refetch();
   };
 
   useEffect(() => {
-    if (usersList) {
-      setCount(usersList.count);
-      setRows([...usersList.records]);
-    }
-  }, [usersList]);
+    setRowsState((prevRowsState) => (data?.records !== undefined ? data.records : prevRowsState));
+  }, [data?.records, setRowsState]);
 
   useEffect(() => {
-    setRowCountState((prevRowCountState) => (count !== undefined ? count : prevRowCountState));
-  }, [count, setRowCountState]);
+    setRowCountState((prevRowCountState) => (data?.count !== undefined ? data.count : prevRowCountState));
+  }, [data?.count, setRowCountState]);
 
   const columns: GridColDef[] = [
     { field: 'id', headerName: 'UID', hide: true, flex: 1 },
@@ -118,43 +92,15 @@ const Users: FunctionComponent<UsersProps> = () => {
       <h1 className="text-4xl font-bold">Usuários</h1>
       <div className="mt-8 w-full h-[74vh]">
         <DataGrid
-          rows={rows}
+          {...dataGridBasePropDefinitions({ isError })}
+          rows={rowsState}
           columns={columns}
-          disableSelectionOnClick
-          pagination
           rowCount={rowCountState}
-          paginationMode="server"
           page={page}
           pageSize={pageSize}
-          rowsPerPageOptions={[10, 15, 20, 30]}
-          loading={isFetching}
+          loading={isLoading || isFetching}
           onPageChange={(newPage) => setPage(newPage)}
           onPageSizeChange={(pageSize) => setPageSize(pageSize)}
-          initialState={{
-            sorting: {
-              sortModel: [{ field: 'createdAt', sort: 'desc' }],
-            },
-          }}
-          sx={{
-            minHeight: '44.5vh',
-            maxHeight: '80vh',
-            borderRadius: 2,
-            '& .MuiCircularProgress-root': {
-              color: '#ef8f01',
-            },
-            '& .MuiCheckbox-root.Mui-checked': {
-              color: '#ef8f01',
-            },
-            '& .MuiDataGrid-row:hover': {
-              backgroundColor: '#ef8f0130',
-            },
-          }}
-          components={{
-            Toolbar: DataGridToolbar,
-            LoadingOverlay: DataGridLoading,
-            NoRowsOverlay: () => <DataGridOverlay error={isError} />,
-          }}
-          localeText={ptBR.components.MuiDataGrid.defaultProps.localeText}
         />
       </div>
       <ConfirmDialog
